feat: Fixed up naming a bit and other general cleanup + added create-tun.sh
The included `create-tun.sh` script handles creating '/dev/net/tun' if it doesn't
already exist, i.e. inside of a docker container. However it is important to note
that in order for this to work inside said container, one must add NET_RAW + NET_ADMIN
capabilities to the container to actually create and manage TUN devices.